+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Added `Wizaplace\SDK\Catalog\Declination::setIsRenewable`
 - Added `Wizaplace\SDK\Catalog\Product::isSubscription`
 - Added `Wizaplace\SDK\Catalog\Product::isRenewable`
+- Added `Wizaplace\SDK\Pim\Option\Option::isSystem`
+- Added `Wizaplace\SDK\Pim\Option\Option::getCode
 
 ## 1.80.0